pyproject.toml
[project]
name = "claudecoach"
version = "0.1.0"
description = "AI-powered running coach — Strava + Claude + Telegram backend."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.92.0",
    "fastapi>=0.135.3",
    "greenlet>=3.4.0",
    "httpx>=0.28.1",
    "jsonschema>=4.26.0",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "sqlalchemy>=2.0.49",
    "uvicorn[standard]>=0.44.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.9",
]
